Update ci.yml
feat: Add Docker image push to GitHub Actions workflow

This commit adds a new step to the existing GitHub Actions workflow that pushes a Docker image to the Docker registry specified in the CI_REGISTRY_IMAGE secret. The image is tagged with the latest version from the semantic-release-dry job. 

The workflow now also ignores markdown files in the push event and has additional tags for release versions.